+++ js/src/main/scala/gpdviz/Frontend.scala
@@ -1,15 +1,12 @@
 package gpdviz
 
 import scala.scalajs.js
-import org.scalajs.dom
 
+// TODO for now, just some webSocket preparations
 object Frontend extends js.JSApp {
 
   def main(): Unit = {
-    // TODO
-    println("Frontend.main: dom.document = " + dom.document)
-    val test = dom.document.getElementById("TEST")
-    println("Frontend.main: test = " + test)
-    dom.document.getElementById("TEST").innerHTML = "SURPI!"
+    new WsListener
   }
+
 }