+++ jvm/src/main/scala/gpdviz/server/GpdvizServer.scala
@@ -3,10 +3,11 @@ package gpdviz.server
 import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
 import akka.stream.ActorMaterializer
-import gpdviz.async.Notifier
+import gpdviz.async.{PusherNotifier, WebSocketNotifier}
 import gpdviz.config.cfg
 import gpdviz.data.{DbInterface, FileDb, PostgresDb}
 
+import scala.concurrent.ExecutionContextExecutor
 import scala.io.StdIn
 
 object GpdvizServer extends GpdvizService {
@@ -14,17 +15,29 @@ object GpdvizServer extends GpdvizService {
     case None     ⇒ new FileDb("data")
     case Some(pg) ⇒ new PostgresDb(pg)
   }
-  val notifier: Notifier = new Notifier
+
+  implicit val system: ActorSystem = ActorSystem()
+  implicit val materializer: ActorMaterializer = ActorMaterializer()
+
+  val (notifier, route) = cfg.pusher match {
+    case Some(pc) ⇒
+      (new PusherNotifier(pc), routes)
+
+    case None     ⇒
+      val wsn = new WebSocketNotifier
+      val wsRoute = path("ws") {
+        handleWebSocketMessages(wsn.wsHandler)
+      }
+      (wsn, routes ~ wsRoute)
+  }
 
   def main(args: Array[String]) {
-    implicit val system = ActorSystem()
-    implicit val materializer = ActorMaterializer()
     // needed for the future flatMap/onComplete in the end
-    implicit val executionContext = system.dispatcher
+    implicit val executionContext: ExecutionContextExecutor = system.dispatcher
 
-    println(s"Gpdviz using: ${db.details}")
+    println(s"Gpdviz using: DB: ${db.details}  Async Notifications: ${notifier.details}")
 
-    val bindingFuture = Http().bindAndHandle(routes, cfg.httpInterface, cfg.httpPort)
+    val bindingFuture = Http().bindAndHandle(route, cfg.httpInterface, cfg.httpPort)
 
     println(s"Gpdviz server '${cfg.serverName}' online at ${cfg.httpInterface}:${cfg.httpPort}/")
     if (!args.contains("-d")) {